partial commit with issues in CatList.ts
